Added Default Device Settings Method and Differential Restore Classes
Added method 'CreateDefaultDevice' to interface 'IBackupRestore'.
The 'CreateDefaultDevice' method was added to the 'BackupBase' and 'RestoreBase' classes to allow you to set the default device.
In the classes 'DifferentialBackup', 'FullBackup', 'TransactionLogBackup', and 'FullRestore', the 'SetDevice' method has been changed to invoke the 'CreateDefaultDevice' method.
A new class of 'Differential Restore' has been added to support differential restore operations of MSSQL databases.