Adding MSSQL.BackupRestore Solutions and Projects
The 'MSSQL.BackupRestore.prototype' project was added to the 'MSSQL.BackupRestore.sln' file.

The build settings for the project 'MSSQL.BackupRestore.Prototype' were added to the 'Global' section of the 'MSSQL.BackupRestore.sln' file.

A new file 'MSSQL.BackupRestore.Prototype.csproj' was added.
This file is targeted at .NET 8.0 and refers to the 'Microsoft.Extensions.Logging.Abstractions' package. It also refers to the 'MSSQL.BackupRestore' project.

A new file 'Program.cs ' has been added, which includes the ability to output logs to the console using the 'ConsoleLoggerFactory' and 'ConsoleLogger' classes.